==> 2_ClusterX_robustness_test.Rmd <==
---
title: "ClusterX Robustness Test"
author: "Hao Chen"
date: "7 June 2016"
output: 
  html_document:
    toc: true
    toc_depth: 2
---


## ClusterX robustness test for alpha selection
```{r}
library(cytofkit)
r15_data <- read.table("./R15_dataset/R15.txt", header = FALSE)[,c(1,2)]

pRange <- seq(0, 0.01, by = 0.001)[-1]
pRange <- c(pRange, 0.02, 0.03, 0.04, 0.05)

cn <- NULL
for(p in pRange){
    c <- ClusterX(r15_data, alpha = p)
    cn <- c(cn, length(unique(c$cluster)))
}

report <- data.frame(p=factor(pRange), clusters = cn)
ggplot(report, aes(x=p, y=clusters)) + 
    geom_bar(colour="black", fill="#DD8888", width=.8, stat="identity") +
    scale_y_continuous(limits = c(0, 20)) +
    theme_bw() + xlab("P value") + ylab("Number of Clusters") +
    theme(axis.title=element_text(size=16,face="bold"))
```

## Plot the cluster results

```{r}
c <- ClusterX(r15_data, alpha = 0.001)
plotData <- cbind(r15_data, cluster = factor(c$cluster))
ggplot(plotData, aes(x=V1, y=V2, colour=cluster)) + geom_point() + theme_bw()
```

## Session Information

```{r}
sessionInfo()
```
